requirements.txt
# =============================================================
# requirements.txt - Python Dependencies
# AI Resume Analyzer
# =============================================================

# Web Framework
flask==3.0.3
werkzeug==3.0.3

# PDF Text Extraction
pdfminer.six==20231228

# NLP - Natural Language Processing
spacy==3.7.5

# Machine Learning - TF-IDF & Cosine Similarity
scikit-learn==1.5.1
numpy==1.26.4
